fix create task
